refactor: Move some From impls to slice files
