+++ src/soa.rs
@@ -1,7 +1,6 @@
-use crate::__alloc::vec::Vec;
 use crate::{
-    AsMutSlice, AsSlice, IntoIter, Iter, IterMut, Slice, SliceMut, SliceRef, SoaClone, SoaRaw,
-    Soars, iter_raw::IterRaw,
+    __alloc::vec::Vec, AsMutSlice, AsSlice, IntoIter, Iter, IterMut, Slice, SliceMut, SliceRef,
+    SoaClone, SoaRaw, Soars, iter_raw::IterRaw,
 };
 use core::{
     borrow::{Borrow, BorrowMut},
@@ -792,46 +791,6 @@ where
     /// Allocate a `Vec<T>` and fill it by moving the contents of `value`.
     fn from(value: Soa<T>) -> Self {
         value.into_iter().collect()
-    }
-}
-
-impl<T> From<&Slice<T>> for Vec<T>
-where
-    T: SoaClone,
-{
-    /// Allocate a `Vec<T>` and fill it by cloning `value`'s items.
-    fn from(value: &Slice<T>) -> Self {
-        value.iter().map(SoaClone::soa_clone).collect()
-    }
-}
-
-impl<T> From<&mut Slice<T>> for Vec<T>
-where
-    T: SoaClone,
-{
-    /// Allocate a `Vec<T>` and fill it by cloning `value`'s items.
-    fn from(value: &mut Slice<T>) -> Self {
-        value.as_ref().into()
-    }
-}
-
-impl<T> From<SliceRef<'_, T>> for Vec<T>
-where
-    T: SoaClone,
-{
-    /// Allocate a `Vec<T>` and fill it by cloning `value`'s items.
-    fn from(value: SliceRef<T>) -> Self {
-        value.as_ref().into()
-    }
-}
-
-impl<T> From<SliceMut<'_, T>> for Vec<T>
-where
-    T: SoaClone,
-{
-    /// Allocate a `Vec<T>` and fill it by cloning `value`'s items.
-    fn from(value: SliceMut<T>) -> Self {
-        value.as_ref().into()
     }
 }
